assignment1/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Flask-Mail
